Add files via upload
This is a simple machine learning and data analysis project of mine in which i have take the titanic data set and analyzed its various features and learnt some insights. The data set is about the people who travelled in the famous Titanic ship which was met with an accident, a century ago. The data set is also filled with the data of whom all are survived and who are not. I have made ML models to work in this and found their accuracy.